+++ src/Field/Data/OptionData.php
@@ -8,13 +8,16 @@ use CraftCms\Cms\Shared\Contracts\Serializable;
 use CraftCms\Cms\Twig\Attributes\AllowedInSandbox;
 use Stringable;
 
-#[AllowedInSandbox]
 class OptionData implements Serializable, Stringable
 {
     public function __construct(
+        #[AllowedInSandbox]
         public ?string $label,
+        #[AllowedInSandbox]
         public ?string $value,
+        #[AllowedInSandbox]
         public bool $selected,
+        #[AllowedInSandbox]
         public bool $valid = true,
         public ?string $icon = null,
         public ?string $color = null,