+++ yii2-adapter/legacy/web/twig/AllowableInSandbox.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license https://craftcms.github.io/license/
+ */
+
+namespace craft\web\twig;
+
+/** @phpstan-ignore-next-line */
+if (false) {
+    /**
+     * Interface AllowableInSandbox
+     *
+     * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+     * @since 4.17.0
+     * @deprecated 6.0.0 use {@see \CraftCms\Cms\Twig\AllowableInSandbox} instead.
+     */
+    interface AllowableInSandbox
+    {
+    }
+}